Fixing the test

+++ commons/utils/job-store-service-connector/src/main/java/dk/dbc/dataio/commons/utils/jobstore/MockedJobStoreServiceConnector.java
@@ -32,7 +32,7 @@ public class MockedJobStoreServiceConnector extends JobStoreServiceConnector {
     }
 
     @Override
-    public JobInfoSnapshot addChunk(Chunk chunk, long jobId, long chunkId) {
+    public JobInfoSnapshot addChunk(Chunk chunk, int jobId, long chunkId) {
         chunks.add(chunk);
         return jobInfoSnapshots.remove();
     }
